backend/graphiti_knowledge_graph_test/Coastline_KG_2/requirements.txt
# Knowledge Graph Dependencies
graphiti-core[falkordb]>=0.24.1
openai>=1.0.0
python-dotenv>=1.0.0

